refactor: use babel to use async/await
