+++ @commitlint/travis-cli/package.json
@@ -3,9 +3,10 @@
   "version": "5.0.1",
   "description": "Lint all relevant commits for a change or PR on Travis CI",
   "bin": {
-    "commitlint-travis": "./cli.js"
+    "commitlint-travis": "./lib/cli.js"
   },
   "scripts": {
+    "build": "npx cross-env NODE_ENV=production npx -p babel-cli babel src --out-dir lib --source-maps",
     "deps": "dep-check",
     "lint": "npx xo",
     "start": "npx ava -c 4 --verbose --watch",
@@ -13,10 +14,15 @@
   },
   "ava": {
     "files": [
-      "*.test.js"
+      "src/*.test.js"
     ],
     "source": [
-      "*.js"
+      "src/*.js"
+    ]
+  },
+  "babel": {
+    "presets": [
+      "commitlint"
     ]
   },
   "xo": false,
@@ -44,6 +50,10 @@
   "devDependencies": {
     "@commitlint/utils": "^5.0.1",
     "ava": "0.18.2",
+    "babel-cli": "6.26.0",
+    "babel-preset-commitlint": "^5.0.1",
+    "babel-register": "6.26.0",
+    "cross-env": "5.1.1",
     "which": "^1.3.0"
   },
   "dependencies": {
